# docs/openapi.generated.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/giving/funds": {
      "get": {
        "operationId": "GivingController_funds",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Liste des fonds",
        "tags": [
          "giving"
        ]
      }
    },
    "/api/v1/giving/payment-methods": {
      "get": {
        "operationId": "GivingController_paymentMethods",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Méthodes de paiement disponibles",
        "tags": [
          "giving"
        ]
      }
    },
    "/api/v1/giving/donations": {
      "post": {
        "operationId": "GivingController_createDonation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDonationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crée un don (initie le paiement)",
        "tags": [
          "giving"
        ]
      },
      "get": {
        "operationId": "GivingController_myDonations",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Champ:direction, ex. createdAt:desc",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Recherche plein texte",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Historique de mes dons",
        "tags": [
          "giving"
        ]
      }
    },
    "/api/v1/giving/donations/{id}": {
      "get": {
        "operationId": "GivingController_getDonation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reçu / détail d’un don",
        "tags": [
          "giving"
        ]
      }
    },
    "/api/v1/giving/webhooks/{provider}": {
      "post": {
        "operationId": "GivingController_webhook",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Webhook provider de paiement (signé)",
        "tags": [
          "giving"
        ]
      }
    },
    "/api/v1/giving/wallet": {
      "get": {
        "operationId": "GivingController_getWallet",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Wallet Amen",
        "tags": [
          "giving"
        ]
      }
    },
    "/api/v1/giving/wallet/topup": {
      "post": {
        "operationId": "GivingController_topup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopupWalletDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Recharge le wallet",
        "tags": [
          "giving"
        ]
      }
    },
    "/api/v1/giving/wallet/default-fund": {
      "patch": {
        "operationId": "GivingController_setDefaultFund",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetDefaultFundDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change le fonds par défaut du wallet",
        "tags": [
          "giving"
        ]
      }
    },
    "/api/v1/giving/wallet/transactions": {
      "get": {
        "operationId": "GivingController_transactions",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Champ:direction, ex. createdAt:desc",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Recherche plein texte",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Transactions du wallet",
        "tags": [
          "giving"
        ]
      }
    },
    "/api/v1/giving/admin/donations": {
      "get": {
        "operationId": "GivingController_adminDonations",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Champ:direction, ex. createdAt:desc",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Recherche plein texte",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fund",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "method",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tous les dons (staff)",
        "tags": [
          "giving"
        ]
      }
    },
    "/api/v1/giving/admin/summary": {
      "get": {
        "operationId": "GivingController_adminSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "KPIs dons (staff)",
        "tags": [
          "giving"
        ]
      }
    },
    "/api/v1/giving/admin/export": {
      "get": {
        "operationId": "GivingController_adminExport",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Export CSV financier (admin)",
        "tags": [
          "giving"
        ]
      }
    },
...
  },
  "info": {
    "title": "IPCK API",
    "description": "API de la plateforme IPCK (mobile + dashboard)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateDonationDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 50,
            "description": "Montant en USD"
          },
          "fundId": {
            "type": "string",
            "example": "general"
          },
          "method": {
            "type": "string",
            "example": "mpesa"
          },
          "anonymous": {
            "type": "boolean",
            "default": false
          },
          "phone": {
            "type": "string",
            "description": "Numéro mobile money (momo)"
          }
        },
        "required": [
          "amount",
          "fundId",
          "method"
        ]
      },
      "WebhookDto": {
        "type": "object",
        "properties": {
          "ref": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "example": "received",
            "enum": [
              "received",
              "failed"
            ]
          },
          "providerRef": {
            "type": "string"
          }
        },
        "required": [
          "ref",
          "status"
        ]
      },
      "TopupWalletDto": {
        "type": "object",
        "properties": {
          "coins": {
            "type": "number",
            "example": 50,
            "description": "Nombre de coins (1 coin = 1 USD)"
          },
          "method": {
            "type": "string",
            "example": "mpesa"
          }
        },
        "required": [
          "coins",
          "method"
        ]
      },
      "SetDefaultFundDto": {
        "type": "object",
        "properties": {
          "fundId": {
            "type": "string",
            "example": "general"
          }
        },
        "required": [
          "fundId"
        ]
      },
...
    }
  }
}